Modulized Admin Screens & refactor view all trips
