+++ mobile-app/Screens/Admin/Admin_View_Users_info.js
@@ -1,0 +1,16 @@
+import {
+    NativeBaseProvider,
+    Text,
+} from 'native-base';
+import React, { useEffect, useState } from 'react';
+
+
+
+
+export const Admin_View_Users_info = ({navigation, route}) => {
+    return (
+        <NativeBaseProvider>
+            <Text>viewing Users info</Text>
+        </NativeBaseProvider>
+    )
+}
